requirements.txt
beautifulsoup4
lxml
pillow
requests[socks]
sklearn
userpass
